➕  서버 설정 UI 추가, 사용되지 않는 참조문 제거 #14

+++ src/main/java/io/github/closeddev/Server/CreateServer.java
@@ -15,10 +15,14 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 import static io.github.closeddev.Main.MCSBPath;
 
 public class CreateServer {
+
+    public static HashMap<String, Object> PROPERTIES = new HashMap<>();
 
     public static final String SERVERPATH = MCSBPath + "/Temp/Testserver";
     public static void createServer(String FullVersion, String Build, String vcode, int ramAmount) {
@@ -28,7 +32,7 @@ public class CreateServer {
             File verjson = new File(MCSBPath + "/versions.json");
             if (!verjson.isFile()) { // version.json 파일이 없을 경우 기본 파일 생성
                 JSONObject jsonObject = new JSONObject();
-                JSONObject versionObj = new JSONObject(); // 이거슨 아무 의미도 없는 JSONObject 이름 원하시면 바꾸시오
+                JSONObject versionObj = new JSONObject();
                 versionObj.put("list", new ArrayList<String>());
                 jsonObject.put("vers", versionObj);
                 JSONManager.writeJSON(MCSBPath + "/versions.json", jsonObject);
@@ -66,9 +70,12 @@ public class CreateServer {
                 vers.put(FullVersion, versionfield);
             }
             JSONManager.writeJSON(MCSBPath + "/versions.json", jsonObject);
-            Files.copy(filename, Paths.get(SERVERPATH + "/paper.jar"));
+            if (!Paths.get(SERVERPATH + "/paper.jar").toFile().exists()) Files.copy(filename, Paths.get(SERVERPATH + "/paper.jar"));
             crteula(SERVERPATH);
             crtstart(SERVERPATH, ramAmount);
+            crtproper();
+            Files.copy(Paths.get((MCSBPath + "/server.properties")), Paths.get(SERVERPATH + "/server.properties"));
+
 
 
         } catch(Exception e) {
@@ -100,8 +107,9 @@ public class CreateServer {
         }
     }
 
-    private static void crtproper(JSONObject properties) {
-        String configPath = Main.PROGRAM_PATH + "/server.properties";
+    public static void crtproper() {
+
+        String configPath = MCSBPath + "/server.properties";
         File config = new File(configPath);
         if (!config.isFile()) {
             try {
@@ -111,6 +119,33 @@ public class CreateServer {
             }
         }
 
-        ServerManager.editProperties(configPath, "enable-command-block", true);
+        for (Map.Entry<String, Object> entry : PROPERTIES.entrySet()) {
+            String key = entry.getKey();
+            Object value = entry.getValue();
+
+            ServerManager.editProperties(configPath, key, value);
+        }
+    }
+
+    public static void setproperdefault() {
+        // Default properties value
+        PROPERTIES.put("level-seed", 0);
+        PROPERTIES.put("gamemode", "survival");
+        PROPERTIES.put("enable-command-block", false);
+        PROPERTIES.put("level-name", "world");
+        PROPERTIES.put("motd", "A Minecraft Server");
+        PROPERTIES.put("difficulty", "easy");
+        PROPERTIES.put("max-players", 20);
+        PROPERTIES.put("online-mode", true);
+        PROPERTIES.put("allow-flight", false);
+        PROPERTIES.put("allow-nether", true);
+        PROPERTIES.put("server-port", 25565);
+        PROPERTIES.put("hardcore", false);
+        PROPERTIES.put("white-list", false);
+        PROPERTIES.put("spawn-npcs", true);
+        PROPERTIES.put("spawn-animals", true);
+        PROPERTIES.put("level-type", "minecraft\\:default");
+        PROPERTIES.put("spawn-monsters", true);
+        PROPERTIES.put("spawn-protection", 16);
     }
 }
